func_remove_flat_curves error
The looping of only selected bones was trowing error. But was working fine when i would have just select one bone. 

I have removed the storing of slected bones and looping only the fcurves related to that.
And storing of fcurves tobe removed and then trying to remove also some how seemed to not work so i am removing the fucurve just after the check if the curve is flat and continuing the loop to check next.

Sorry if this is not a right way of doing it. but with my limited knowledge this was make do solution